+++ CHANGELOG.md
@@ -14,6 +14,8 @@ List of major changes and improvements
 
   - `Network` uses `None` in previous/next layers for the input/output layers.
 
+  - `Network` methods to return the first/last layers.
+
   - Add batch size argument to layer fmap size methods.
 
   - Add default filter size to `FCLayer`.